+++ src/contentTree/layerContentTreeItem.js
@@ -9,11 +9,11 @@ import component from '../actions/styleSelector.vue';
 import VcsObjectContentTreeItem from './vcsObjectContentTreeItem.js';
 import { vcsAppSymbol } from '../pluginHelper.js';
 import { contentTreeClassRegistry } from './contentTreeItem.js';
+import { executeCallbacks } from '../callback/vcsCallback.js';
 
 /**
  * @typedef {ContentTreeItemOptions} LayerContentTreeItemOptions
  * @property {string} layerName
- * @property {Array<string>} layerNamesToDeactivate list of LayerNames which should be deactivated if the click activates the layer
  */
 
 /**
@@ -85,14 +85,6 @@ class LayerContentTreeItem extends VcsObjectContentTreeItem {
      * @private
      */
     this._layerName = options.layerName;
-
-    /**
-     * @type {Array<string>}
-     * @private
-     */
-    this._layerNamesToDeactivate = Array.isArray(options.layerNamesToDeactivate)
-      ? options.layerNamesToDeactivate.slice()
-      : [];
 
     /**
      * @type {Array<Function>}
@@ -238,15 +230,14 @@ class LayerContentTreeItem extends VcsObjectContentTreeItem {
   }
 
   async clicked() {
+    await super.clicked();
     if (this._layer) {
       if (this.state === StateActionState.INACTIVE) {
         await this._layer.activate();
-        this._layerNamesToDeactivate
-          .map((n) => this._app.layers.getByKey(n))
-          .filter((l) => l)
-          .forEach((l) => l.deactivate());
+        executeCallbacks(this._app, this._onActivate);
       } else {
         this._layer.deactivate();
+        executeCallbacks(this._app, this._onDeactivate);
       }
     }
   }
@@ -257,7 +248,6 @@ class LayerContentTreeItem extends VcsObjectContentTreeItem {
   toJSON() {
     const config = super.toJSON();
     config.layerName = this._layerName;
-    config.layerNamesToDeactivate = this._layerNamesToDeactivate.slice();
     return config;
   }