+++ tests/unit/contentTree/layerGroupContentTreeItem.spec.js
@@ -30,7 +30,6 @@ describe('LayerGroupContentTreeItem', () => {
     let app;
     let item;
     let layers;
-    let layerToDeactivate;
 
     beforeEach(async () => {
       app = new VcsUiApp();
@@ -46,13 +45,10 @@ describe('LayerGroupContentTreeItem', () => {
       layers.forEach((l) => {
         app.layers.add(l);
       });
-      layerToDeactivate = new VectorLayer({ name: 'layerToDeactivate' });
-      app.layers.add(layerToDeactivate);
       item = new LayerGroupContentTreeItem(
         {
           name: 'foo',
           layerNames: layers.map((l) => l.name),
-          layerNamesToDeactivate: [layerToDeactivate.name],
         },
         app,
       );
@@ -118,12 +114,6 @@ describe('LayerGroupContentTreeItem', () => {
         expect(layers.filter((l) => l.active)).to.have.members(layers);
       });
 
-      it('should deactivate all layers given by layersToDeactivate', async () => {
-        await layerToDeactivate.activate();
-        await item.clicked();
-        expect(layerToDeactivate.active).to.be.false;
-      });
-
       it('should activate all inactive layers, if some layers are active', async () => {
         const [layer1] = layers;
         await layer1.activate();
@@ -165,7 +155,6 @@ describe('LayerGroupContentTreeItem', () => {
         inputConfig = {
           name: 'foo',
           layerNames: ['foo'],
-          layerNamesToDeactivate: ['foo2'],
         };
 
         const item = new LayerGroupContentTreeItem(inputConfig, app);
@@ -173,20 +162,8 @@ describe('LayerGroupContentTreeItem', () => {
         item.destroy();
       });
 
-      it('should only configure type, name, layerNames and layerNamesToDeactivate', () => {
-        expect(outputConfig).to.have.all.keys([
-          'name',
-          'type',
-          'layerNames',
-          'layerNamesToDeactivate',
-        ]);
-      });
-
-      it('should configure layerNamesToDeactivate', () => {
-        expect(outputConfig).to.have.property(
-          'layerNamesToDeactivate',
-          outputConfig.layerNamesToDeactivate,
-        );
+      it('should only configure type, name and layerNames', () => {
+        expect(outputConfig).to.have.all.keys(['name', 'type', 'layerNames']);
       });
 
       it('should configure layerNames', () => {
@@ -207,7 +184,6 @@ describe('LayerGroupContentTreeItem', () => {
         inputConfig = {
           name: 'foo',
           layerNames: ['foo'],
-          layerNamesToDeactivate: ['foo2'],
           defaultViewpoint: 'foo',
         };